Exercise2.1: Blogs can only be deleted by users who created them

+++ blog/controller/blogs.js
@@ -31,6 +31,21 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  const blogUserId = blog.user
+
+  if (blogUserId.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'unathorized' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
